Conflicts:
	Documentation/filesystems/overlayfs.txt
	arch/arm64/configs/bonito_defconfig
	drivers/android/binder.c
	drivers/char/adsprpc.c
	drivers/cpuidle/lpm-levels.c
	drivers/input/keydebug-core.c
	drivers/input/keydebug-func.c
	drivers/media/platform/msm/vidc/msm_vidc_common.c
	drivers/power/supply/overheat_mitigation.c
	drivers/staging/android/ion/ion.c
	fs/f2fs/data.c
	fs/f2fs/segment.c
	fs/f2fs/super.c
	fs/overlayfs/super.c
	include/linux/mm_event.h
	include/trace/events/mm_event.h
	include/uapi/linux/android/binder.h
	mm/Kconfig
	mm/filemap.c
	mm/mm_event.c

Bug: 119540871

+++ drivers/net/usb/lan78xx.c
@@ -869,7 +869,7 @@ static int lan78xx_read_otp(struct lan78xx_net *dev, u32 offset,
 
 	if (ret == 0) {
 		if (sig == OTP_INDICATOR_1)
-			offset = offset;
+			offset = (u32)offset;
 		else if (sig == OTP_INDICATOR_2)
 			offset += 0x100;
 		else
